move tests from test_numeric.py to test_summarize.py
